Created Inventory System, Fixed Scene Properties
A new Inventory System is implemented. Items can be picked up on mouse click and stored in a Scroll View. Items can be deleted from inventory. Current issue is that an extra item is being generated in the Inventory, which needs to be tackled in future updates. The scene was also slightly modified to reduce any non realistic collisions with the player.